fix installing completions with sdist
Instead, let's just switch to using package_data instead of data_files.
Add an extra command for installing completions so we don't have to deal
with all the python permission junk. This keeps all data neatly in the
password-sanity directory. Also bump the version up to 0.2 to fix this
on pypi.