pws: use an encrypt_master function
There was some silly duplicated logic when encrypting the new file. Put
this in a function. Also the behavior is changed to be a little safer.
When writing a new master file, it is written to a temporary path first
and as long as the encryption/writing succeeds, then the old file is
removed and the new file gets renamed. This, in theory, would avoid a
bad/corrupted partial write that must quit and throw an error. In the
old code, this partial write would have gone directly over the master.asc
which would be quite bad.