pws: change clipboard-copy syntax
Technically a breaking change. I naively assumed that every OS would
have a normal clipboard copy that goes something like "command --args
$value", but of course I was wrong. On Windows, it seems like you must
specify the string in question in the middle of the command if you don't
want to grab an extra newline and pipe it to clip. Because of this, now
users (implying anyone uses this) now will need to include a "$value" in
the clipboard-copy variable. So what was once "xclip" is now just "xclip
$value".