pws: escape characters more properly
Sending strings to a shell containing special characters is a problem of
course. I was being lazy previously and just wrapping it in '' so the
entire quote as treated as a quote literal (and also omitting ' from the
list of characters used to generatea password). This was crap so instead
just escape every single character in the string before fusing it the
clipboard command and sending it to the shell for execution. There's
also some windows-specific logic (oh no) in here because the escape
character for cmd is "^" not "\". There are some more windows-related
changes that will come in the next commit.